Added IsFollowed to ProfileJSON, replaced many UserIds with type int64

+++ api/internal/server/websocket/WSClients.go
@@ -11,7 +11,7 @@ type ClientList map[*Client]bool
 
 type Client struct {
 	connection *websocket.Conn
-	clientID   int
+	clientID   int64
 	manager    *WebsocketServer
 	gate       chan Payload
 }
@@ -28,7 +28,7 @@ var (
 	maxMessageSize int64 = 512
 )
 
-func NewClient(conn *websocket.Conn, userID int, manager *WebsocketServer) *Client {
+func NewClient(conn *websocket.Conn, userID int64, manager *WebsocketServer) *Client {
 	return &Client{
 		connection: conn,
 		clientID:   userID,
