+++ api/pkg/models/post.go
@@ -10,7 +10,7 @@ import (
 
 type Post struct {
 	Id          int
-	UserId      int
+	UserId      int64
 	Content     string
 	ImagePath   string
 	CreatedAt   time.Time
@@ -30,7 +30,7 @@ type FeedPost struct {
 
 type IPostRepository interface {
 	GetAllByUserId(id int64) ([]*Post, error)
-	GetAllFeedPosts(currentUserId int, offset int) ([]*FeedPost, error)
+	GetAllFeedPosts(currentUserId int64, offset int) ([]*FeedPost, error)
 	GetById(id int64) (*Post, error)
 	Insert(post *Post) (int64, error)
 	GetCommentCount(postId int) (int, error)
@@ -122,7 +122,7 @@ func (repo PostRepository) GetAllByUserId(id int64) ([]*Post, error) {
 }
 
 // Return all posts to the current user by offset
-func (m PostRepository) GetAllFeedPosts(currentUserId int, offset int) ([]*FeedPost, error) {
+func (m PostRepository) GetAllFeedPosts(currentUserId int64, offset int) ([]*FeedPost, error) {
 
 	//Change value if needed for testing purposes
 	// currentUserId = 11
